+++ aider/repo.py
@@ -8,6 +8,7 @@ try:
     ANY_GIT_ERROR = [
         git.exc.ODBError,
         git.exc.GitError,
+        git.exc.InvalidGitRepositoryError,
     ]
 except ImportError:
     git = None
@@ -26,6 +27,8 @@ ANY_GIT_ERROR += [
     BufferError,
     TypeError,
     ValueError,
+    AttributeError,
+    AssertionError,
 ]
 ANY_GIT_ERROR = tuple(ANY_GIT_ERROR)
 
@@ -285,9 +288,17 @@ class GitRepo:
                 files = self.tree_files[commit]
             else:
                 try:
-                    for blob in commit.tree.traverse():
-                        if blob.type == "blob":  # blob is a file
-                            files.add(blob.path)
+                    iterator = commit.tree.traverse()
+                    while True:
+                        try:
+                            blob = next(iterator)
+                            if blob.type == "blob":  # blob is a file
+                                files.add(blob.path)
+                        except IndexError:
+                            self.io.tool_warning(f"GitRepo: read error skipping {blob.path}")
+                            continue
+                        except StopIteration:
+                            break
                 except ANY_GIT_ERROR as err:
                     self.git_repo_error = err
                     self.io.tool_error(f"Unable to list files in git repo: {err}")
@@ -359,8 +370,8 @@ class GitRepo:
 
     def ignored_file_raw(self, fname):
         if self.subtree_only:
-            fname_path = Path(self.normalize_path(fname))
             try:
+                fname_path = Path(self.normalize_path(fname))
                 cwd_path = Path.cwd().resolve().relative_to(Path(self.root).resolve())
             except ValueError:
                 # Issue #1524